all working

+++ Survive The Night/Assets/Game/Scripts/EndOfRoundController.cs
@@ -286,9 +286,9 @@ public class EndOfRoundController : MonoBehaviour
         }
         else
         {
-            int previousHealth = _gameController.getWallHealth();
+            double previousHealth = _gameController.getWallHealth();
             _gameController.RepairBase(_repairsHoursSelectedValue * 3);
-            int newHealth = _gameController.getWallHealth();
+            double newHealth = _gameController.getWallHealth();
 
             _summaryController.updateBaseRepairsText("Your base was repaired from " + previousHealth + "% -> " + newHealth + "%");
         }
